requirements.txt
Django==1.9.5
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-classy-tags==0.6.2
django-datetime-widget==0.9.3
django-easy-maps==0.9.2
django-imagekit==3.3
django-swingtime==0.6.2
geopy==1.11.0
html5lib==0.999
httplib2==0.8
lxml==3.3.3
mutagen==1.22
MySQL-python==1.2.5
oauthlib==0.6.1
oneconf==0.3.7.14.4.1
PAM==0.4.2
pexpect==3.1
pilkit==1.1.13
Pillow==2.3.0
piston-mini-client==0.7.5
psutil==1.2.1
pycrypto==2.6.1
pycups==1.9.66
pygobject==3.12.0
pyOpenSSL==0.13
pyserial==2.6
pysmbc==1.0.14.1
python-apt===0.9.3.5ubuntu2
python-dateutil==2.4.2
python-debian==0.1.23
pytz==2015.7
pyxdg==0.25
reportlab==3.0
requests==2.2.1
sessioninstaller==0.0.0
six==1.10.0
software-center-aptd-plugins==0.0.0
Twisted-Core==13.2.0
Twisted-Web==13.2.0
urllib3==1.7.1
zope.interface==4.0.5
